refactor course introduction comp
